requirements.txt
requests
google-api-python-client==2.48.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
tenacity==8.0.1
python-magic
requests
python-dotenv
bs4
pyrogram==2.0.30